Tasco Cood by id function viewGood in GoodController and view row fileds in View tascogood

+++ database/migrations/2020_11_08_141133_create_goods_table.php
@@ -35,7 +35,7 @@ class CreateGoodsTable extends Migration
             $table->string('descriptionkg');
             $table->string('descriptiontr');
             $table->string('descriptiontc');
-            $table->integer('avgstarscount');
+            $table->double('avgstarscount');
             $table->integer('commentcount');
             $table->integer('ordercount');
             $table->double('price');
